# TS29222_CAPIF_Security_API/openapi.yaml
openapi: 3.0.0
info:
  description: |
    API for CAPIF security management.
    © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  title: CAPIF_Security_API
  version: 1.2.0-alpha.1
...
servers:
- url: '{apiRoot}/capif-security/v1'
  variables:
    apiRoot:
      default: https://example.com
      description: apiRoot as defined in subclause 7.5 of 3GPP TS 29.222.
paths:
  /trustedInvokers/{apiInvokerId}:
    delete:
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: apiInvokerId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: No Content (Successful deletion of the existing subscription)
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "308":
          description: Permanent Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
    get:
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: apiInvokerId
        required: true
        schema:
          type: string
        style: simple
      - description: When set to 'true', it indicates the CAPIF core function to send
          the authentication information of the API invoker. Set to false or omitted
          otherwise.
        explode: true
        in: query
        name: authenticationInfo
        required: false
        schema:
          type: boolean
        style: form
      - description: When set to 'true', it indicates the CAPIF core function to send
          the authorization information of the API invoker. Set to false or omitted
          otherwise.
        explode: true
        in: query
        name: authorizationInfo
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceSecurity'
          description: The security related information of the API Invoker based on
            the request from the API exposing function.
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "308":
          description: Permanent Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Acceptable
        "414":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: URI Too Long
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
    put:
      callbacks:
        notificationDestination:
          '{request.body#/notificationDestination}':
            post:
              operationId: notificationDestination_request_bodyNotificationDestinationPost
              requestBody:
                content:
                  application/json:
                    schema:
                      $ref: '#/components/schemas/SecurityNotification'
                required: true
              responses:
                "204":
                  description: No Content (successful notification)
                "307":
                  description: Temporary Redirect
                  headers:
                    Location:
                      description: An alternative URI of the resource.
                      required: true
                      schema:
                        type: string
                "308":
                  description: Permanent Redirect
                  headers:
                    Location:
                      description: An alternative URI of the resource.
                      required: true
                      schema:
                        type: string
                "400":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Bad request
                "401":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Unauthorized
                "403":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Forbidden
                "404":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Not Found
                "411":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Length Required
                "413":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Payload Too Large
                "415":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Unsupported Media Type
                "429":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Too Many Requests
                "500":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Internal Server Error
                "503":
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                  description: Service Unavailable
                default:
                  description: Generic Error
              x-callback-request: true
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: apiInvokerId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceSecurity'
        description: create a security context for an API invoker
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceSecurity'
          description: Successful created.
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according
                to the structure: {apiRoot}/capif-security/v1/trustedInvokers/{apiInvokerId}'
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "411":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Length Required
        "413":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Payload Too Large
        "414":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: URI Too Long
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unsupported Media Type
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
  /trustedInvokers/{apiInvokerId}/update:
    post:
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: apiInvokerId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceSecurity'
        description: Update the security context (e.g. re-negotiate the security methods).
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceSecurity'
          description: Successful updated.
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "308":
          description: Permanent Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "411":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Length Required
        "413":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Payload Too Large
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unsupported Media Type
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
  /trustedInvokers/{apiInvokerId}/delete:
    post:
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: apiInvokerId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityNotification'
        description: Revoke the authorization of the API invoker for APIs.
        required: true
      responses:
        "204":
          description: Successful revoked.
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "308":
          description: Permanent Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "411":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Length Required
        "413":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Payload Too Large
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unsupported Media Type
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
  /securities/{securityId}/token:
    post:
      parameters:
      - description: Identifier of an individual API invoker
        explode: false
        in: path
        name: securityId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AccessTokenReq'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenRsp'
          description: Successful Access Token Request
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "308":
          description: Permanent Redirect
          headers:
            Location:
              description: An alternative URI of the resource.
              required: true
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenErr'
          description: Error in the Access Token Request
components:
  schemas:
    ServiceSecurity:
      description: Represents the details of the security method for each service
        API interface. When included by the API invoker, it indicates the preferred
        method of security. When included by the CAPIF core function, it indicates
        the security method to be used for the service API interface.
      example:
        notificationDestination: notificationDestination
        supportedFeatures: supportedFeatures
        securityInfo:
        - authenticationInfo: authenticationInfo
          authorizationInfo: authorizationInfo
          interfaceDetails:
            ipv6Addr: ipv6Addr
            securityMethods:
            - null
            - null
            port: 5248
            ipv4Addr: ipv4Addr
          prefSecurityMethods:
          - null
          - null
          aefId: aefId
        - authenticationInfo: authenticationInfo
          authorizationInfo: authorizationInfo
          interfaceDetails:
            ipv6Addr: ipv6Addr
            securityMethods:
            - null
            - null
            port: 5248
            ipv4Addr: ipv4Addr
          prefSecurityMethods:
          - null
          - null
          aefId: aefId
        websockNotifConfig:
          requestWebsocketUri: true
          websocketUri: websocketUri
        requestTestNotification: true
      properties:
        securityInfo:
          items:
            $ref: '#/components/schemas/SecurityInformation'
          minimum: 1
          title: securityInfo
          type: array
        notificationDestination:
          description: string providing an URI formatted according to IETF RFC 3986.
          title: notificationDestination
          type: string
        requestTestNotification:
          description: Set to true by API invoker to request the CAPIF core function
            to send a test notification as defined in in subclause 7.6. Set to false
            or omitted otherwise.
          title: requestTestNotification
          type: boolean
        websockNotifConfig:
          $ref: '#/components/schemas/WebsockNotifConfig'
        supportedFeatures:
          description: A string used to indicate the features supported by an API
            that is used as defined in clause 6.6 in 3GPP TS 29.500. The string shall
            contain a bitmask indicating supported features in hexadecimal representation
            Each character in the string shall take a value of "0" to "9", "a" to
            "f" or "A" to "F" and shall represent the support of 4 features as described
            in table 5.2.2-3. The most significant character representing the highest-numbered
            features shall appear first in the string, and the character representing
            features 1 to 4 shall appear last in the string. The list of features
            and their numbering (starting with 1) are defined separately for each
            API. If the string contains a lower number of characters than there are
            defined features for an API, all features that would be represented by
            characters that are not present in the string are not supported
          pattern: ^[A-Fa-f0-9]*$
          title: supportedFeatures
          type: string
      required:
      - notificationDestination
      - securityInfo
      title: ServiceSecurity
      type: object
...
    SecurityNotification:
      description: Represents revoked authorization notification details.
      example:
        apiInvokerId: apiInvokerId
        aefId: aefId
        apiIds:
        - apiIds
        - apiIds
      properties:
        apiInvokerId:
          description: String identifying the API invoker assigned by the CAPIF core
            function.
          type: string
        aefId:
          description: String identifying the AEF.
          type: string
        apiIds:
          description: Identifier of the service API
          items:
            type: string
          minItems: 1
          type: array
        cause:
          $ref: '#/components/schemas/Cause'
      required:
      - apiIds
      - apiInvokerId
      - cause
      type: object
    AccessTokenReq:
      description: Represents the access token request information.
      format: x-www-form-urlencoded
      properties:
        grant_type:
          enum:
          - client_credentials
          type: string
        client_id:
          type: string
        client_secret:
          type: string
        scope:
          type: string
      required:
      - client_id
      - grant_type
...
    ProblemDetails:
      description: Represents additional information and details on an error response.
      properties:
        type:
          description: string providing an URI formatted according to IETF RFC 3986.
          title: notificationDestination
          type: string
        title:
          description: A short, human-readable summary of the problem type. It should
            not change from occurrence to occurrence of the problem.
          title: title
          type: string
        status:
          description: The HTTP status code for this occurrence of the problem.
          title: status
          type: integer
        detail:
          description: A human-readable explanation specific to this occurrence of
            the problem.
          title: detail
          type: string
        instance:
          description: string providing an URI formatted according to IETF RFC 3986.
          title: notificationDestination
          type: string
        cause:
          description: A machine-readable application error cause specific to this
            occurrence of the problem. This IE should be present and provide application-related
            error information, if available.
          title: cause
          type: string
        invalidParams:
          description: Description of invalid parameters, for a request rejected due
            to invalid parameters.
          items:
            $ref: '#/components/schemas/InvalidParam'
          minItems: 1
          title: invalidParams
          type: array
        supportedFeatures:
          description: A string used to indicate the features supported by an API
            that is used as defined in clause 6.6 in 3GPP TS 29.500. The string shall
            contain a bitmask indicating supported features in hexadecimal representation
            Each character in the string shall take a value of "0" to "9", "a" to
            "f" or "A" to "F" and shall represent the support of 4 features as described
            in table 5.2.2-3. The most significant character representing the highest-numbered
            features shall appear first in the string, and the character representing
            features 1 to 4 shall appear last in the string. The list of features
            and their numbering (starting with 1) are defined separately for each
            API. If the string contains a lower number of characters than there are
            defined features for an API, all features that would be represented by
            characters that are not present in the string are not supported
          pattern: ^[A-Fa-f0-9]*$
          title: supportedFeatures
          type: string
      title: ProblemDetails
      type: object
...